# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "APIResponse_DatasetResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DatasetResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Endpoint specific payload."
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[DatasetResponse]",
        "type": "object"
      },
      "APIResponse_DatasetUploadResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DatasetUploadResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Endpoint specific payload."
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[DatasetUploadResponse]",
        "type": "object"
      },
...
      "APIResponse_LabelingJobResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LabelingJobResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Endpoint specific payload."
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[LabelingJobResponse]",
        "type": "object"
      },
      "APIResponse_NoneType_": {
        "properties": {
          "data": {
            "description": "Endpoint specific payload.",
            "title": "Data",
            "type": "null"
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[NoneType]",
        "type": "object"
      },
...
      "APIResponse_list_DatasetResponse__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DatasetResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Endpoint specific payload.",
            "title": "Data"
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[list[DatasetResponse]]",
        "type": "object"
      },
...
      "Body_upload_dataset_api_datasets__dataset_id__upload_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          },
          "replace_existing": {
            "default": false,
            "title": "Replace Existing",
            "type": "boolean"
          },
          "text_column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Column"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_dataset_api_datasets__dataset_id__upload_post",
        "type": "object"
      },
      "DatasetCreate": {
        "description": "Payload used to create a dataset.",
        "properties": {
          "dataset_type": {
            "maxLength": 50,
            "minLength": 2,
            "title": "Dataset Type",
            "type": "string"
          },
          "description": {
            "maxLength": 5000,
            "minLength": 10,
            "title": "Description",
            "type": "string"
          },
          "title": {
            "maxLength": 150,
            "minLength": 3,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "description",
          "dataset_type"
        ],
        "title": "DatasetCreate",
        "type": "object"
      },
...
      "DatasetResponse": {
        "description": "Dataset metadata returned to the frontend.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dataset_type": {
            "maxLength": 50,
            "minLength": 2,
            "title": "Dataset Type",
            "type": "string"
          },
          "description": {
            "maxLength": 5000,
            "minLength": 10,
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "record_count": {
            "title": "Record Count",
            "type": "integer"
          },
          "source_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Filename"
          },
          "source_format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Format"
          },
          "source_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Size Bytes"
          },
          "title": {
            "maxLength": 150,
            "minLength": 3,
            "title": "Title",
            "type": "string"
          },
          "uploaded_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded At"
          }
        },
        "required": [
          "title",
          "description",
          "dataset_type",
          "id",
          "owner_id",
          "source_filename",
          "source_format",
          "source_size_bytes",
          "record_count",
          "uploaded_at",
          "created_at"
        ],
        "title": "DatasetResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LabelingJobUpdate": {
        "description": "Payload used to update a labeling job.",
        "properties": {
          "annotation_type": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annotation Type"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 5000,
                "minLength": 10,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "label_options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 30,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label Options"
          },
          "status": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 150,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "LabelingJobUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend API for the AI Dataset Labeling Marketplace: datasets, labeling jobs, annotation tasks, annotations, quality review, progress tracking, dataset export and the AI annotation assistant.",
    "title": "AI Dataset Labeling Marketplace API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/datasets": {
      "get": {
        "description": "Return datasets visible to the authenticated role.",
        "operationId": "get_datasets_api_datasets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_DatasetResponse__"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Datasets",
        "tags": [
          "Datasets"
        ]
      },
      "post": {
        "description": "Create a dataset container owned by the authenticated user.",
        "operationId": "create_dataset_api_datasets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_DatasetResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Dataset",
        "tags": [
          "Datasets"
        ]
      }
    },
    "/api/datasets/{dataset_id}": {
      "delete": {
        "description": "Delete a dataset and its dependent workflow records.",
        "operationId": "delete_dataset_api_datasets__dataset_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Dataset",
        "tags": [
          "Datasets"
        ]
      },
      "get": {
        "description": "Return one dataset when it exists and is visible to the caller.",
        "operationId": "get_dataset_api_datasets__dataset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_DatasetResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Dataset",
        "tags": [
          "Datasets"
        ]
      },
      "put": {
        "description": "Update dataset metadata after checking ownership.",
        "operationId": "update_dataset_api_datasets__dataset_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_DatasetResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Dataset",
        "tags": [
          "Datasets"
        ]
      }
    },
    "/api/datasets/{dataset_id}/upload": {
      "post": {
        "description": "Import CSV, JSON, JSONL or XLSX records into the dataset.",
        "operationId": "upload_dataset_api_datasets__dataset_id__upload_post",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_dataset_api_datasets__dataset_id__upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_DatasetUploadResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload and parse a dataset file",
        "tags": [
          "Datasets"
        ]
      }
    },
    "/api/exports/datasets/{dataset_id}": {
      "get": {
        "description": "Download the approved labels of a whole dataset.",
        "operationId": "export_dataset_api_exports_datasets__dataset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "csv",
              "pattern": "^(csv|json)$",
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_unapproved",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Unapproved",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export the labeled records of every job in a dataset",
        "tags": [
          "Exports"
        ]
      }
    },
    "/api/exports/jobs/{job_id}": {
      "get": {
        "description": "Download the approved labels of a job as CSV or JSON.",
        "operationId": "export_job_api_exports_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "csv",
              "pattern": "^(csv|json)$",
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_unapproved",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Unapproved",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export the labeled records of a labeling job",
        "tags": [
          "Exports"
        ]
      }
    },
...
    "/api/jobs/{job_id}": {
      "delete": {
        "description": "Delete a labeling job and its dependent work items.",
        "operationId": "delete_job_api_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete a labeling job",
        "tags": [
          "Labeling Jobs"
        ]
      },
      "get": {
        "description": "Return a single labeling job by identifier.",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_LabelingJobResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retrieve one labeling job",
        "tags": [
          "Labeling Jobs"
        ]
      },
      "put": {
        "description": "Update the title, description or status of a job.",
        "operationId": "update_job_api_jobs__job_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelingJobUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_LabelingJobResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update a labeling job",
        "tags": [
          "Labeling Jobs"
        ]
      }
    },
...
  }
}